Refactored to more closely represent ECS Architecture
