+++ core/src/main/java/io/github/alltheeb5t/unisim/GameScreen.java
@@ -20,55 +20,30 @@ import com.badlogic.gdx.utils.viewport.Viewport;
 
 import io.github.alltheeb5t.unisim.map_objects.MapBuilding;
 
-public class GameScreen extends ScreenAdapter implements InputProcessor {
+public class GameScreen extends ScreenAdapter {
 
     private OrthographicCamera camera;
     private SpriteBatch batch;
     private World world;
     private Stage stage;
-    private Viewport viewport;
     private Box2DDebugRenderer box2dDebugRenderer;
-    private MapBuilding testBuilding;
 
-    private DragAndDrop dragAndDrop;
-
-    // Used by to track how much the cursor has moved between two 'touchDragged' events.
-    // The camera is then translated by this.
-    private int panStartX;
-    private int panStartY;
+    private CampusMap campusMap;
 
     public GameScreen (OrthographicCamera camera, Viewport viewport) {
         this.camera = camera;
-        this.viewport = viewport;
 
         batch = new SpriteBatch();
         world = new World(new Vector2(0, 0), false);
-        box2dDebugRenderer = new Box2DDebugRenderer();
         stage = new Stage(viewport);
-        dragAndDrop = new DragAndDrop();
-        Gdx.input.setInputProcessor(this); // Inputs related to drag are manually passed to stage
 
-        // Testing only: Add temporary building
-        testBuilding = new MapBuilding(480, 100, 120, world, new Texture("piazza.png"));
-        stage.addActor(testBuilding);
+        box2dDebugRenderer = new Box2DDebugRenderer();
 
-        // Drag and drop listener for test building
-        dragAndDrop.addSource(new DragAndDrop.Source(testBuilding) {
-            @Override
-			public DragAndDrop.Payload dragStart(InputEvent inputEvent, float v, float v1, int i) {
-                DragAndDrop.Payload payload = new DragAndDrop.Payload();
-				payload.setDragActor(getActor());
-                stage.addActor(testBuilding);
-				dragAndDrop.setDragActorPosition(getActor().getWidth() / 2, -getActor().getHeight() / 2);
+        // Custom Logic Component Instantiations
+        campusMap = new CampusMap(camera, stage, world);
 
-				return payload;
-			}
-
-			@Override
-			public void dragStop(InputEvent event, float x, float y, int pointer, DragAndDrop.Payload payload, DragAndDrop.Target target) {
-				System.out.println("Drag stop");
-			}
-        });
+        Gdx.input.setInputProcessor(campusMap.getInputSystem()); // Inputs related to drag are manually passed to stage
+        
     }
 
     @Override
@@ -86,114 +61,9 @@ public class GameScreen extends ScreenAdapter implements InputProcessor {
 
     }
 
-    /** 
-     * Moves the camera by a set amount but prevents going beyond map boundaries
-     */
-    private void clampedTranslate(float x, float y) {
-        camera.position.x += x;
-        camera.position.y += y;
-
-		float effectiveViewportWidth = camera.viewportWidth * camera.zoom;
-		float effectiveViewportHeight = camera.viewportHeight * camera.zoom;
-
-        // Ensure that the camera's viewport never goes beyond the edge of the map
-        camera.position.x = MathUtils.clamp(camera.position.x, effectiveViewportWidth / 2f, Constants.CAMPUS_MAX_X - effectiveViewportWidth / 2f);
-        camera.position.y = MathUtils.clamp(camera.position.y, effectiveViewportHeight / 2f, Constants.CAMPUS_MAX_Y - effectiveViewportHeight / 2f);
-		
-	}
-
     @Override
     public void resize(int width, int height) {
-        stage.getViewport().update(width, height);
-        camera.viewportWidth = width;
-        camera.viewportHeight = height;
-        camera.update();
-    }
-
-    @Override
-    public boolean mouseMoved(int screenX, int screenY) {
-        return false;
-    }
-
-    @Override
-    public boolean touchDown(int screenX, int screenY, int pointer, int button) {
-        panStartX = screenX;
-        panStartY = screenY;
-        return stage.touchDown(screenX, screenY, pointer, button);
-    }
-
-    /**
-     * Implemented from InputProcessor. Used to handle panning if middle click is pressed
-     * @param screenX Pointer X coordinate
-     * @param screenY Pointer Y coordinate
-     * @param pointer Think this is always 0 for a mouse.
-     * @return
-     */
-    @Override
-    public boolean touchDragged (int screenX, int screenY, int pointer) {
-        if (Gdx.input.isButtonPressed(Input.Buttons.MIDDLE))  {
-            clampedTranslate((panStartX-screenX), -(panStartY-screenY));
-            panStartX = screenX;
-            panStartY = screenY;
-        }
-        
-        // Input handler would ordinarily be set to the stage. Now we've finished, pass input events to stage.
-        return stage.touchDragged(screenX, screenY, pointer);
-    }
-
-    @Override
-    public boolean touchUp (int screenX, int screenY, int pointer, int button) {
-        return stage.touchUp(screenX, screenY, pointer, button);
-    }
-
-	@Override public boolean keyDown (int keycode) {
-		return false;
-	}
-
-	@Override public boolean keyUp (int keycode) {
-		return false;
-	}
-
-	@Override public boolean keyTyped (char character) {
-		return false;
-	}
-
-    /**
-     * Implemented from InputProcessor, used to control Zoom.
-     * @param amountX Arbitrary value representing how far the window should move.
-     * @param amountY Would be horizontal scrolling. Not used.
-     * @return
-     */
-	@Override public boolean scrolled (float amountX, float amountY) {
-        // We find the in-game coordinates of the cursor initially
-        Vector3 cursorPosMeters = camera.unproject(new Vector3(Gdx.input.getX(), Gdx.input.getY(), 0));
-        camera.zoom += amountY/10;
-
-        // Minimum zoom is the entire viewport filling the screen and the maximum is an arbitrary 0.1 (On a 1080p monitor this would be ~98m)
-		camera.zoom = MathUtils.clamp(camera.zoom, 0.1f, Math.min(Constants.CAMPUS_MAX_X/camera.viewportWidth, Constants.CAMPUS_MAX_Y/camera.viewportHeight));
-
-        camera.update();
-
-        // Get the in-game coordinates of the cursor now the zoom has been executed
-        Vector3 newCursorPosMeters = camera.unproject(new Vector3(Gdx.input.getX(), Gdx.input.getY(), 0));
-        
-        // To make zoom natural, whatever is directly under the cursor should stay in the same place
-        clampedTranslate(cursorPosMeters.x-newCursorPosMeters.x,cursorPosMeters.y-newCursorPosMeters.y);
-        camera.update();
-		return true;
-	}
-
-    /**
-     * Implemented as is necessary for a full implementation of InputProcessor. Not relevant on desktop
-     * @param screenX
-     * @param screenY
-     * @param point
-     * @param button
-     * @return
-     */
-    @Override
-    public boolean touchCancelled(int screenX, int screenY, int pointer, int button) {
-        return false;
+        campusMap.getInputSystem().resize(width, height);
     }
